refactor(Locale): Change format of Locale Enum
