+++ src/main/scala/com/particeep/api/models/enums/Locale.scala
@@ -1,9 +1,8 @@
 package com.particeep.api.models.enums
 
-object Locale {
+sealed trait Locale extends Product with Enum
 
-  sealed abstract class Locale extends Product with Enum
-
+object Locale extends EnumHelper[Locale] {
   case object fr_FR extends Locale
   case object en_GB extends Locale
   case object en_US extends Locale
@@ -15,7 +14,5 @@ object Locale {
   case object nl_BE extends Locale
   case object ar_AR extends Locale
 
-  object Locale extends EnumHelper[Locale] {
-    def values: Set[Locale] = Set(fr_FR, en_GB, en_US, es_ES, de_DE, pt_PT, pt_BR, nl_NL, nl_BE, ar_AR)
-  }
+  def values: Set[Locale] = Set(fr_FR, en_GB, en_US, es_ES, de_DE, pt_PT, pt_BR, nl_NL, nl_BE, ar_AR)
 }
